django/requirements.txt
# Django
Django==4.2
django-environ
gunicorn
django-debug-toolbar

# PostgreSQL
psycopg2

# MySQL
# TODO: Better to create seperate files for diffrent database.
mysqlclient

# Celery
celery
redis
django-celery-beat

# DRF
djangorestframework
django-filter
drf-writable-nested
djangorestframework_simplejwt
django-cors-headers
